changed castle music a bit and did a few other things

+++ src/game/camera.h
@@ -479,7 +479,7 @@ struct CameraFOVStatus {
     /// How much to decrease shakeAmplitude each frame.
     /*0x18*/ s16 decay;
 
-    u16 multiplier;
+    u8 multiplier;
     f32 velocity;
     f32 inc;
 };
